requirements.txt
streamlit
pymupdf
stable-baselines3
gymnasium
llama-cpp-python
